move EloqModule related symbols to eloq_module.h
