+++ src/bthread/bthread.cpp
@@ -30,42 +30,6 @@
 
 DECLARE_bool(use_io_uring);
 
-extern std::array<eloq::EloqModule *, 10> registered_modules;
-extern std::atomic<int> registered_module_cnt;
-
-int register_module(eloq::EloqModule *module) {
-    static std::mutex module_mutex;
-    std::unique_lock<std::mutex> lk(module_mutex);
-    size_t i = 0;
-    while (i < registered_modules.size() && registered_modules[i] != nullptr) {
-        // Each module should only be registered once.
-        CHECK(registered_modules[i] != module);
-        i++;
-    }
-    registered_modules[i] = module;
-    registered_module_cnt.fetch_add(1, std::memory_order_release);
-    return 0;
-}
-
-int unregister_module(eloq::EloqModule *module) {
-    static std::mutex module_mutex;
-    std::unique_lock<std::mutex> lk(module_mutex);
-    size_t i = 0;
-    while (i < registered_modules.size() && registered_modules[i] != module) {
-        i++;
-    }
-    if (i == registered_modules.size()) {
-        return 0;
-    }
-    CHECK(i < registered_module_cnt);
-    while (i < registered_modules.size() - 1) {
-        registered_modules[i] = registered_modules[i+1];
-        i++;
-    }
-    registered_module_cnt.fetch_sub(1, std::memory_order_release);
-    return 0;
-}
-
 
 namespace bthread {
